requirements.txt
asgiref==3.2.10
astroid==2.2.5
autopep8==1.4.4
certifi==2020.6.20
colorama==0.4.1
Django==3.0
djangorestframework==3.11.1
isort==4.3.17
lazy-object-proxy==1.3.1
mccabe==0.6.1
pycodestyle==2.5.0
pylint==2.3.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
typed-ast==1.3.4
wincertstore==0.2
wrapt==1.11.1
